CH-53E: Merged Pierre Duval's work.
This is a merge of Pierre's excellent model and changes.

All I've done is to merge, rework the sounds, add a splash screen, changed the folding so that it doesn't matter if the blades are in the right position they will stil end up folded, although they may pass through parts of the aircraft to do so. I also started on the ALS cockpit effects (panel backlighting and interior shadows)

There are a lot of files that have been reorganised (e.g. Nasal) and as the Pierre's file movements fit with the way models are usually done I've kept his layout rather than merging his changes over the top of the existing model (which would be better for version control history but that's all really).

Some of the models will be in the next commit (due to svn problem with next-id) 

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/ch53e@949 9eb9bc84-db9d-4e05-893b-26dd41ac94a6